+++ exonum-java-binding/core/rust/src/cmd/run.rs
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-use exonum_cli::command::{run::Run as StandardRun, ExonumCommand, StandardResult};
-use failure;
+use anyhow;
+use exonum_cli::command::{ExonumCommand, Run as StandardRun, StandardResult};
 use serde::{Deserialize, Serialize};
 use structopt::StructOpt;
 
@@ -67,7 +67,7 @@ pub struct Run {
 }
 
 impl EjbCommand for Run {
-    fn execute(self) -> Result<EjbCommandResult, failure::Error> {
+    fn execute(self) -> Result<EjbCommandResult, anyhow::Error> {
         if let StandardResult::Run(node_run_config) = self.standard.execute()? {
             let jvm_config = JvmConfig {
                 args_prepend: self.jvm_args_prepend,
@@ -91,12 +91,12 @@ impl EjbCommand for Run {
             };
 
             let config = Config {
-                run_config: node_run_config,
+                run_config: *node_run_config,
                 jvm_config,
                 runtime_config,
             };
 
-            Ok(EjbCommandResult::EjbRun(config))
+            Ok(EjbCommandResult::EjbRun(Box::new(config)))
         } else {
             unreachable!("Standard run command returned invalid result")
         }
@@ -105,7 +105,7 @@ impl EjbCommand for Run {
 
 /// Returns full path to the default log configuration file assuming the `exonum-java` app is
 /// packaged/installed.
-fn get_path_to_default_log_config() -> PathBuf {
+pub(crate) fn get_path_to_default_log_config() -> PathBuf {
     let mut path = executable_directory();
     path.push("log4j-fallback.xml");
     path