+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/service/ExecutionContext.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2019 The Exonum Team
+ * Copyright 2020 The Exonum Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.core.transaction;
+package com.exonum.binding.core.service;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -26,13 +26,21 @@ import com.exonum.binding.core.blockchain.Blockchain;
 import com.exonum.binding.core.blockchain.BlockchainData;
 import com.exonum.binding.core.runtime.ServiceInstanceSpec;
 import com.exonum.binding.core.storage.database.Prefixed;
+import com.exonum.binding.core.transaction.Transaction;
+import java.util.Optional;
+import javax.annotation.Nullable;
 
 /**
- * Transaction context class. Contains required information for the transaction execution.
- * The context is provided by the framework and users shouldn't create context instances manually
- * except tests.
+ * An execution context. The context provides access to the blockchain data
+ * for the executing service, and also contains the required information for the transaction
+ * execution.
+ *
+ * <p>The context is provided by the framework and users shouldn't create context instances manually
+ * except in tests.
+ *
+ * @see Transaction
  */
-public interface TransactionContext {
+public interface ExecutionContext {
 
   /**
    * Returns the prefixed database access for the executing service. Allows R/W operations.
@@ -54,18 +62,22 @@ public interface TransactionContext {
 
   /**
    * Returns SHA-256 hash of the {@linkplain TransactionMessage transaction message} that
-   * carried the payload of the transaction.
-   * Each transaction message is uniquely identified by its hash; the messages are persisted
+   * carried the payload of the transaction; or {@code Optional.empty()} if no message corresponds
+   * to this context.
+   *
+   * <p>Each transaction message is uniquely identified by its hash; the messages are persisted
    * in the {@linkplain Blockchain#getTxMessages() blockchain} and can be fetched by this hash.
    */
-  HashCode getTransactionMessageHash();
+  Optional<HashCode> getTransactionMessageHash();
 
   /**
-   * Returns public key of the transaction author. The corresponding transaction message
-   * is guaranteed to have a correct {@link CryptoFunctions#ed25519()} signature
-   * with this public key.
+   * Returns public key of the transaction author; or {@code Optional.empty()} if no transaction
+   * message corresponds to this context.
+   *
+   * <p>The corresponding transaction message, if any, is guaranteed to have a correct
+   * {@link CryptoFunctions#ed25519()} signature with this public key.
    */
-  PublicKey getAuthorPk();
+  Optional<PublicKey> getAuthorPk();
 
   /**
    * Returns the name of the service instance.
@@ -109,7 +121,7 @@ public interface TransactionContext {
     /**
      * Sets transaction message hash for the context.
      */
-    public Builder txMessageHash(HashCode hash) {
+    public Builder txMessageHash(@Nullable HashCode hash) {
       this.hash = hash;
       return this;
     }
@@ -117,7 +129,7 @@ public interface TransactionContext {
     /**
      * Sets transaction author public key for the context.
      */
-    public Builder authorPk(PublicKey authorPk) {
+    public Builder authorPk(@Nullable PublicKey authorPk) {
       this.authorPk = authorPk;
       return this;
     }
@@ -141,8 +153,8 @@ public interface TransactionContext {
     /**
      * Creates the transaction context instance.
      */
-    public TransactionContext build() {
-      return InternalTransactionContext.newInstance(blockchainData, hash, authorPk, serviceName,
+    public ExecutionContext build() {
+      return InternalExecutionContext.newInstance(blockchainData, hash, authorPk, serviceName,
           checkNotNull(serviceId));
     }