+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/service/Service.java
@@ -16,8 +16,6 @@
 
 package com.exonum.binding.core.service;
 
-import com.exonum.binding.core.blockchain.BlockchainData;
-import com.exonum.binding.core.transaction.ExecutionException;
 import io.vertx.ext.web.Router;
 
 /**
@@ -40,17 +38,16 @@ public interface Service {
    * or save all or some configuration parameters as is for later retrieval in transactions
    * and/or read requests.
    *
-   * @param blockchainData blockchain data accessor for this service. Not valid after this method
-   *     returns
+   * @param context the execution context
    * @param configuration the service configuration parameters
    * @throws ExecutionException if the configuration parameters are not valid (e.g.,
    *     malformed, or do not meet the preconditions). Exonum will stop the service if
    *     its initialization fails. It will save the error into
-   *     {@linkplain com.exonum.binding.core.blockchain.Blockchain#getCallErrors(long)
+   *     {@linkplain com.exonum.binding.core.blockchain.Blockchain#getCallRecords(long)}
    *     the registry of call errors}
    * @see Configurable
    */
-  default void initialize(BlockchainData blockchainData, Configuration configuration) {
+  default void initialize(ExecutionContext context, Configuration configuration) {
     // No configuration
   }
 
@@ -66,13 +63,12 @@ public interface Service {
    * when the block is committed.
    * <!--TODO: Add a link to the migration procedure -->
    *
-   * @param blockchainData blockchain data accessor for this service. Not valid after this method
-   *     returns
+   * @param context the execution context
    * @param arguments the service arguments
    * @throws ExecutionException if the arguments are not valid (e.g.,
    *     malformed, or do not meet the preconditions)
    */
-  default void resume(BlockchainData blockchainData, byte[] arguments) {
+  default void resume(ExecutionContext context, byte[] arguments) {
     // No actions by default
   }
 
@@ -108,12 +104,12 @@ public interface Service {
 
   /**
    * An optional callback method invoked by the blockchain <em>before</em> any transactions
-   * in a block are executed. See {@link #afterTransactions(BlockchainData)} for details.
+   * in a block are executed. See {@link #afterTransactions(ExecutionContext)} for details.
    *
-   * @see #afterTransactions(BlockchainData)
+   * @see #afterTransactions(ExecutionContext)
    * @see com.exonum.binding.core.transaction.Transaction
    */
-  default void beforeTransactions(BlockchainData blockchainData) {}
+  default void beforeTransactions(ExecutionContext context) {}
 
   /**
    * Handles the changes made by all transactions included in the upcoming block.
@@ -128,18 +124,17 @@ public interface Service {
    *
    * <p>Any exceptions in this method will revert any changes made to the database by it,
    * but will not affect the processing of this block. Exceptions are saved
-   * in {@linkplain com.exonum.binding.core.blockchain.Blockchain#getCallErrors(long)
+   * in {@linkplain com.exonum.binding.core.blockchain.Blockchain#getCallRecords(long)
    * the registry of call errors} with appropriate error kinds.
    *
-   * @param blockchainData blockchain data accessor for this service. Not valid after this method
-   *     returns
+   * @param context the execution context
    * @throws ExecutionException if an error occurs during the method execution;
    *     it is saved as a call error of kind "service". Any other exceptions
    *     are considered unexpected. They are saved with kind "unexpected".
-   * @see #beforeTransactions(BlockchainData)
+   * @see #beforeTransactions(ExecutionContext)
    * @see com.exonum.binding.core.transaction.Transaction
    */
-  default void afterTransactions(BlockchainData blockchainData) {}
+  default void afterTransactions(ExecutionContext context) {}
 
   /**
    * Handles read-only block commit event. This handler is an optional callback method which is