+++ exonum-java-binding/testkit/src/main/java/com/exonum/binding/testkit/TestKit.java
@@ -39,12 +39,14 @@ import com.exonum.binding.core.runtime.FrameworkModule;
 import com.exonum.binding.core.runtime.ServiceRuntimeAdapter;
 import com.exonum.binding.core.service.BlockCommittedEvent;
 import com.exonum.binding.core.service.Configuration;
+import com.exonum.binding.core.service.ExecutionContext;
 import com.exonum.binding.core.service.Node;
 import com.exonum.binding.core.service.Service;
 import com.exonum.binding.core.storage.database.Prefixed;
 import com.exonum.binding.core.storage.database.Snapshot;
 import com.exonum.binding.core.storage.indices.KeySetIndexProxy;
 import com.exonum.binding.core.storage.indices.MapIndex;
+import com.exonum.binding.core.testkit.internal.TestKitProtos;
 import com.exonum.binding.core.testkit.internal.TestKitProtos.TestKitServiceInstances;
 import com.exonum.binding.core.transaction.RawTransaction;
 import com.exonum.binding.core.transport.Server;
@@ -66,7 +68,6 @@ import io.vertx.ext.web.Router;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -86,8 +87,8 @@ import javax.annotation.Nullable;
  * from the pool are committed when a new block is created with {@link #createBlock()}.
  *
  * <p>When TestKit is created, Exonum blockchain instance is initialized — service instances are
- * {@linkplain Service#initialize(com.exonum.binding.core.blockchain.BlockchainData, Configuration)
- * initialized} and genesis block is committed.
+ * {@linkplain Service#initialize(ExecutionContext,
+ * Configuration) initialized} and genesis block is committed.
  * Then the {@linkplain Service#createPublicApiHandlers(Node, Router) public API handlers} are
  * created.
  *
@@ -588,23 +589,22 @@ public final class TestKit extends AbstractCloseableNativeProxy {
       checkDeployedArtifactsAreUsed();
       TestKitServiceInstances.Builder builder = TestKitServiceInstances.newBuilder();
 
-      // Add specifications of artifacts to deploy.
-      for (Map.Entry<ServiceArtifactId, String> entry : serviceArtifactFilenames.entrySet()) {
-        ServiceArtifactId artifactId = entry.getKey();
+      for (ServiceArtifactId artifactId : services.keySet()) {
+        String artifactFilename = serviceArtifactFilenames.get(artifactId);
         Base.ArtifactSpec artifactSpec = Base.ArtifactSpec.newBuilder()
             .setArtifact(artifactIdToProto(artifactId))
             .setPayload(DeployArguments.newBuilder()
-                .setArtifactFilename(entry.getValue())
+                .setArtifactFilename(artifactFilename)
                 .build()
                 .toByteString())
             .build();
 
-        builder.addArtifactSpecs(artifactSpec);
-      }
+        TestKitProtos.TestKitService.Builder serviceBuilder =
+            TestKitProtos.TestKitService.newBuilder();
+        serviceBuilder.setArtifactSpec(artifactSpec);
+        serviceBuilder.addAllServiceSpecs(services.get(artifactId));
 
-      // Add specifications of service instances to start.
-      for (Lifecycle.InstanceInitParams instanceInitParams : services.values()) {
-        builder.addServiceSpecs(instanceInitParams);
+        builder.addServices(serviceBuilder.build());
       }
 
       return builder.build();